requirements.txt
Django==3.0.5
djangorestframework==3.11.0
mysqlclient
gunicorn==20.0.4
gevent==1.4.0
uvicorn
